+++ fr.lip6.move.gal.itscl.application/src/fr/lip6/move/gal/itscl/application/Application.java
@@ -6,14 +6,11 @@ import java.util.ArrayList;
 import java.util.Hashtable;
 import java.util.List;
 
+import org.eclipse.emf.common.util.EList;
 import org.eclipse.equinox.app.IApplication;
 import org.eclipse.equinox.app.IApplicationContext;
 
-import fr.lip6.move.gal.CTLProp;
-import fr.lip6.move.gal.LTLProp;
-import fr.lip6.move.gal.Property;
-import fr.lip6.move.gal.SafetyProp;
-import fr.lip6.move.gal.Specification;
+
 import fr.lip6.move.gal.instantiate.GALRewriter;
 import fr.lip6.move.gal.itstools.CommandLine;
 import fr.lip6.move.gal.itstools.CommandLineBuilder;
@@ -28,8 +25,7 @@ public class Application implements IApplication {
 	private static final String CTL_EXAM = "-ctl";
 	private static final String LTL_EXAM = "-ltl";
 	
-	private Thread th1;
-	
+
 	/* (non-Javadoc)
 	 * @see org.eclipse.equinox.app.IApplication#start(org.eclipse.equinox.app.IApplicationContext)
 	 */
@@ -95,22 +91,19 @@ public class Application implements IApplication {
 		Specification spec = SerializationUtil.fileToGalSystem(inputff);		
 		System.out.println("Successfully read input file : " + inputff +" in " + (time - System.currentTimeMillis()) + " ms.");
 		
-		
-		
 		time = System.currentTimeMillis();
 		Problem p = loadModel(pwd, spec, tool);
 		System.out.println("Simplifications done in " + (time - System.currentTimeMillis()) + " ms.");
 		
 		time = System.currentTimeMillis();
-		CommandLine cl= getCmdLine(spec,p.getFolder(),modelName,tool,p);
+		CommandLine cl= getCmdLine(spec,p.getFolder(),modelName,tool);
 		System.out.println("Built GAL and property files in "+ (time - System.currentTimeMillis()) + " ms.");
-		SolverObservable obs= new SolverObservable();
-		SolverSeq s = (SolverSeq) new ItsSolver(p,cl,obs);
-		obs.attach(s);
-		th1=s.solve();
 		
-
-	
+		SolverObservable obs= new SolverObservable();
+		SolverSeq s = new SolverSeq(p,cl);
+		obs.attach(s);
+		s.solve(obs);
+		
 		return IApplication.EXIT_OK;
 	}
 	
@@ -128,15 +121,15 @@ public class Application implements IApplication {
 		}		
 		
 		GALRewriter.flatten(spec, true);
-		return new Problem(spec,tool,3500,cwd);
+		return new Problem(spec,tool,3500,cwd,spec.getProperties());
 	}
 	
 	
 	// On produit un fichier de modèle pour l'outil ligne de commande
-	public CommandLine getCmdLine(Specification spec, String cwd, String modelName,Tool tool,Problem pr) throws IOException{
+	public CommandLine getCmdLine(Specification spec, String cwd, String modelName,Tool tool) throws IOException{
 		
 		String outpath = cwd+"/"+ modelName + ".gal";
-		pr.outputGalFile(spec, outpath);	
+		outputGalFile(spec, outpath);	
 		
 		CommandLine cl = buildCommandLine(outpath, tool);
 	
@@ -170,6 +163,17 @@ public class Application implements IApplication {
 
 	}
 		
+	
+	public void outputGalFile(Specification spec, String outpath) throws IOException {
+		if (! spec.getProperties().isEmpty()) {
+			List<Property> props = new ArrayList<Property>(spec.getProperties());
+			spec.getProperties().clear();
+			SerializationUtil.systemToFile(spec, outpath);
+			spec.getProperties().addAll(props);
+		} else {
+			SerializationUtil.systemToFile(spec, outpath);
+		}
+	}
 	
 	private CommandLine buildCommandLine(String modelff, Tool tool) throws IOException {
 		CommandLineBuilder cl = new CommandLineBuilder(tool);
